feat: modularize codebase by consolidating duplicated code
- Remove duplicated SVGSelection class from pstricks package, use only @latex2js/utils version
- Move working arrow function from NextJS app to packages/pstricks/src/lib/psgraph.ts
- Move working psgraph const from NextJS app to packages/pstricks/src/lib/psgraph.ts
- Update NextJS app to import arrow and psgraph from @latex2js/pstricks package
- Export arrow function from pstricks package index
- Fix TypeScript error in NextJS app page.tsx
- All packages build successfully and NextJS app functions correctly
