+++ packages/pstricks/src/index.ts
@@ -1,7 +1,8 @@
 import pstricks from './lib/pstricks';
-import psgraph from './lib/psgraph';
-export { pstricks, psgraph };
+import psgraph, { arrow } from './lib/psgraph';
+export { pstricks, psgraph, arrow };
 export default {
   pstricks,
   psgraph,
+  arrow,
 };
